:recycle: Feat: meetingSchdule로 로직 이동
<body>
미팅에 대해 정산이 존재해 로직 위치 이동

- 관련 : #343
