+++ application/external-api-v2/src/main/java/com/namo/spring/application/external/api/schedule/controller/MeetingScheduleController.java
@@ -172,4 +172,18 @@ public class MeetingScheduleController {
         return ResponseDto.onSuccess("모임 나가기 성공");
     }
 
+    @Operation(summary = "스케줄 정산내역 (전체 정산) API", description = "스케줄의 모든 활동에 대한 전체 정산 조회하기")
+    @ApiErrorCodes(value = {
+            NOT_FOUND_PARTICIPANT_FAILURE,
+    })
+    @GetMapping("/{scheduleId}/settlement")
+    public ResponseDto<MeetingScheduleResponse.ScheduleSettlementDto> getScheduleSettlement(
+            @AuthenticationPrincipal SecurityUserDetails memberInfo,
+            @Parameter(description = "전체 정산할 스케줄(scheduleId) ID 입니다.", example = "1")
+            @PathVariable Long scheduleId
+    ){
+        return ResponseDto.onSuccess(meetingScheduleUsecase
+                .getScheduleSettlement(memberInfo.getUserId(), scheduleId));
+    }
+
 }
