+++ application/external-api-v2/src/main/java/com/namo/spring/application/external/api/schedule/controller/ScheduleController.java
@@ -52,18 +52,4 @@ public class ScheduleController {
         scheduleUsecase.updateOrCreateScheduleReminder(dto, scheduleId, member);
         return ResponseDto.onSuccess("알림 수정 성공");
     }
-
-    @Operation(summary = "스케줄 정산내역 (전체 정산) API", description = "스케줄의 모든 활동에 대한 전체 정산 조회하기")
-    @ApiErrorCodes(value = {
-            NOT_FOUND_PARTICIPANT_FAILURE,
-    })
-    @GetMapping("/{scheduleId}/settlement")
-    public ResponseDto<MeetingScheduleResponse.ScheduleSettlementDto> getScheduleSettlement(
-            @AuthenticationPrincipal SecurityUserDetails memberInfo,
-            @Parameter(description = "전체 정산할 스케줄(scheduleId) ID 입니다.", example = "1")
-            @PathVariable Long scheduleId
-    ){
-        return ResponseDto.onSuccess(scheduleUsecase
-                .getScheduleSettlement(memberInfo.getUserId(), scheduleId));
-    }
 }
